requirements.txt
django
pillow
django-rest-framework
cx_Oracle